requirements.txt
langchain-ollama
langchain-experimental
langchain-huggingface
langchain-core
langchain-community
langchain
torch
pyspark
nltk
rank_bm25
contractions
unicodedata2
joblib
streamlit
neo4j
python-dotenv
sentence-transformers
faiss-cpu
pandas
regex